==> 3. API dataset with label for clean or not clean (cleanornot_all_makeup_products_API.csv).ipynb <==
import pandas as pd

all_makeup_csv_file = "all_product_details.csv" 
df_makeup = pd.read_csv(all_makeup_csv_file)

clean_makeup_csv_file = "clean_makeup_product_details.csv" 
df_clean_makeup = pd.read_csv(clean_makeup_csv_file)


# strip any leading/trailing spaces from column names
df_makeup.columns = df_makeup.columns.str.strip()
df_clean_makeup.columns = df_clean_makeup.columns.str.strip()


clean_product_ids = set(df_clean_makeup['Product ID'])

#new binary column 'is_clean'
df_makeup['is_clean'] = df_makeup['Product ID'].apply(lambda x: 1 if x in clean_product_ids else 0)


print("First few rows of the updated 'All Makeup' DataFrame:")
print(df_makeup[['Product ID', 'Product Name', 'is_clean']].head())

# Saving the updated DataFrame to a new CSV file
updated_makeup_csv_file = "cleanornot_all_makeup_products_API.csv"
df_makeup.to_csv(updated_makeup_csv_file, index=False)

print(f"Updated 'All Makeup' data with 'is_clean' to {updated_makeup_csv_file}")
