==> 4. Merging the unclean cleanornot API makeup with unclean cleanornot kaggle makeup.ipynb <==
import pandas as pd\

kaggledf = pd.read_csv('makeup_kaggle.csv')  # Adjust path as needed
APIdf = pd.read_csv('cleanornot_all_makeup_products_API.csv')  # Adjust path as needed

# Merging the datasets on 'product_id' from kaggledf and 'Product ID' from APIdf
merged_df = pd.merge(kaggledf, APIdf, left_on='product_id', right_on='Product ID', how='inner')

merged_df.to_csv('unclean_merged_dataset.csv', index=False)  # Save to a new file

pd.set_option('display.max_columns', None)
print(merged_df.columns) 

print(merged_df.shape)


#save kaggle all makeup to csv
makeup_csv_file = "unclean_cleanornot_APIandKaggle.csv"
merged_df.to_csv(makeup_csv_file, index=False)

print(f"makeup data has been saved to {makeup_csv_file}")
